// templates/io-assist-angular-editable/src/app/io-assist/lib/shared/enums/loading-state.enum.ts
export enum LOADING_STATE {
    LOADING = "loading",
    SUCCESS = "success",
    NOT_STARTED = "not_started",
    ERROR = "error",
}

export enum MESSAGES_LOADING_STATE {
    FETCH_MESSAGES_FROM_THREAD_SUCCESS = "FETCH_MESSAGES_FROM_THREAD_SUCCESS",
    GET_RESPONSE_SUCCESS = "GET_RESPONSE_SUCCESS",
    ERROR = "ERROR",
    LOADING_FROM_THREAD = "LOADING_FROM_THREAD",
    LOADING_RESPONSE = "LOADING_RESPONSE",
    NOT_STARTED = "NOT_STARTED",
}